fix the test cases
